+++ src/main/java/constants/Constants.java
@@ -2,6 +2,8 @@ package constants;
 
 public class Constants {
 	public static final int RANDOM_NUMBER_SIZE = 3;
+	public static final int RESTART_NUMBER_SIZE = 1;
+
 	public static final int START_NUMBER = 1;
 	public static final int END_NUMBER = 9;
